requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.1
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==10.6.0
aiosmtplib==3.0.1
apscheduler==3.10.4
email-validator==2.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
